chore(bindings): clippy fixes

+++ bindings/cdk-js/src/types/amount.rs
@@ -36,7 +36,7 @@ impl JsAmount {
     #[wasm_bindgen(constructor)]
     pub fn new(sats: u64) -> Self {
         Self {
-            inner: Amount::from(sats as u64),
+            inner: Amount::from(sats),
         }
     }
